day13

+++ day13/src/oop/inherit4/Mp4.java
@@ -2,14 +2,6 @@ package oop.inherit4;
 
 public class Mp4 extends MediaFile {
 	private float speed;
-
-	public float getSpeed() {
-		return speed;
-	}
-
-	public void setSpeed(int speed) {
-		this.speed = speed;
-	}
-	
-	
+	public float getSpeed() { return speed; }
+	public void setSpeed(float speed) { this.speed = speed; }
 }
